requirements.txt
python-nmap
tabulate
loguru 